requirements.txt
aiohttp==3.8.3
discord.py==2.1.0
requests==2.27.1
youtube_dl==2021.12.17
